requirements.txt
# 项目依赖
python-telegram-bot==20.7
pyyaml==6.0.1
psutil==5.9.6
smartmontools==0.3
paramiko==3.4.0
schedule==1.2.1
watchdog==3.0.0
